refactor(updater): declare the IPC verbs once, dispatch the impl per platform
The Windows split declared all four updater commands twice — once in
updater_macos.rs, once in updater_stub.rs — which broke
tests/ipc-contract.test.ts. That guard reads source text and cannot
evaluate #[cfg], so two mutually exclusive declarations look exactly like
a genuine handler collision, and Tauri does key its IPC router on the bare
fn name. Loosening the guard would have cost more than it bought.

The signatures had already drifted, too: the stub's update_ignore was a
sync fn(version) while the real one is an async fn(version, app), and the
stub omitted the injected AppHandle/State arguments everywhere. Only one
side compiles per target, so nothing caught it.

So the four #[tauri::command] wrappers and the two DTOs they return now
live once in updater.rs — which is also what CONTRIBUTING.md asks for
("IPC verbs grouped into a single commands/<domain>.rs"). The platform
modules expose pub(super) fns behind identical signatures, so a change to
one that isn't mirrored in the other now fails to compile on that platform
instead of drifting.

updater_macos.rs stays byte-identical to the previous updater.rs apart
from those mechanical edits: the two DTO definitions move to the parent,
`use super::{UpdateStatus, UpdaterSnapshot}` comes in, and the four
`#[tauri::command] pub` become `pub(super)`. No logic changes.

Also drops the dispatcher's rustdoc links to `updater_macos`/`updater_stub`
— both load as `mod imp`, so neither name resolves — and documents why the
stub's update_apply returns Err rather than a silent Ok.
